# Reuse Arc allocation if the ref count is one