@orium 2018-02-22
Yeah, the speedup can be quite drastic if you have no structural sharing (or not a lot of it).  However I would like to keep the API nice and clean.

My idea is to have a `VectorMut` with the mutable methods and `Vector` would just be a small wrapper around it, where all its methods would be immutable except for a `pub fn mutable(&mut self) -> &mut VectorMut<T>`.  This way you can choose to:

1. Use only `Vector<T>` and have a nice functional way of using the data structure.  Performance will not be as good.
2. Use only `VectorMut<T>` for the best performance.  Cloning is cheap if you need to keep previous versions of the data structure.
3. Use `Vector<T>` but when you don't need to keep previous versions you can call `vector.mutable().push_back()` to get the best performance.