@Marwes 2018-02-23
Is it really worth it to create a separate type just to keep the interface of the "immutable only" types small? The "immutable only" type has nothing preventing them from supporting the methods and there is not that many methods that needs a mutable variant either. For instance, Vector` needs `push_back`, `set`, ``drop_last` which is only 2 methods more than adding a `mutable` method.

Anyway, I made this mostly out of curiosity of how low the overhead would be with this strategy :) I can clean this `Vector` implementation up and finish the other methods if you want it but otherwise feel free to close or pick apart whatever parts are useful.