@orium 2018-02-24
> Is it really worth it to create a separate type just to keep the interface of the "immutable only" types small?

It is more about having the interface consistent, but I haven't made my mind yet.  For now I'm happy with merging this after a cleanup.