@Marwes 2018-03-09
Implemented `drop_last_mut` as well as a few assorted that fell out of mutability (or I just noticed them as I went along).