@orium 2018-03-14
The changes made by rustfmt are still present, but it gave me the oportunity to rustfmt the code in master (I just landed that).  So, before you rebase create a `.rustfmt.toml` in the project root with

```
write_mode = "Diff"

comment_width = 100
struct_field_align_threshold = 4
use_field_init_shorthand = true
use_try_shorthand = true
match_arm_blocks = false
```

then do

```
cargo +nightly fmt -- --write-mode=overwrite
```

After this commit the changes and you should be able to rebase from master cleanly.