@orium 2018-04-01
Looks good!  I have only a few minor changes, but since I will splitting this in `Vector`/`VectorMut` (see https://github.com/orium/rpds/tree/mutable) those changes will get ironed out then.

For now I'm happy to land this after a new rebase on master and getting those commits squashed into a single one.